Connect searching with the HTTP API

+++ semantic_search/http.py
@@ -3,7 +3,7 @@ from dataclasses import asdict
 from aiohttp import web
 
 from semantic_search.config import Config
-from semantic_search.elastic import ElasticSearchGateway
+from semantic_search.elastic import ElasticSearchGateway, IndexNotAvailable
 from semantic_search.model import Embedder
 from semantic_search.wikipedia import scrap_from_titles_file
 
@@ -12,10 +12,13 @@ async def search(request):
     elastic: ElasticSearchGateway = request.app['elastic']
     search_query = request.query['query']
 
-    result = await elastic.search(search_query)
+    try:
+        results = await elastic.search(search_query)
+    except IndexNotAvailable:
+        return web.Response(status=503,
+                            body="Elasticsearch index not available. Please POST /reindex")
 
-    result_serialized = [asdict(entry) for entry in result]
-    return web.json_response(data=result_serialized, status=200)
+    return web.json_response(data=results, status=200)
 
 
 async def reindex(request):
@@ -25,7 +28,8 @@ async def reindex(request):
     titles_file = Config.TITLES_FILE_PATH
     articles = scrap_from_titles_file(titles_file)
 
-    await elastic.remove_index()
+    if await elastic.index_available():
+        await elastic.remove_index()
     await elastic.create_index()
     for article in articles:
         article_with_embeddings = await embedder.embed_article(article)
